pyproject.toml
[tool.poetry]
name = "weatherapp"
version = "0.1.0"
description = "Weather web app on Python (Django)"
authors = ["Nikita Shibaev <nshib.54321@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.10"
django = "^5.2.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
